Fix PG column type and handling of big IDs in code (MODHAADM-21) (#45)
- The Postgres column for harvestable ID must be `bigint` to hold the new long numbers.
  - The code should not use Integer functions or variables but rather Long everywhere when handling harvestableIds

+++ src/main/java/org/folio/harvesteradmin/moduledata/HarvestJobField.java
@@ -2,7 +2,7 @@ package org.folio.harvesteradmin.moduledata;
 
 public enum HarvestJobField implements Field {
   ID("id", "id", PgColumn.Type.UUID, false, true, true),
-  HARVESTABLE_ID("harvestableId", "harvestable_id", PgColumn.Type.INTEGER, false, true),
+  HARVESTABLE_ID("harvestableId", "harvestable_id", PgColumn.Type.BIGINT, false, true),
   HARVESTABLE_NAME("name", "harvestable_name", PgColumn.Type.TEXT, false, true),
   HARVESTABLE_TYPE("type", "type", PgColumn.Type.TEXT, false, true),
   URL("url", "url", PgColumn.Type.TEXT, false, false),
